m2-openapi - splitted up entities internal models to separate projects and packages
