--- ok-workoutapp-common/src/commonMain/kotlin/models/WktEquipment.kt
@@ -1,16 +1,0 @@
-package models
-
-enum class WktEquipment(s: String) {
-  NONE("None"),
-  OWN_WEIGHT("OwnWeight"),
-  DUMBBELLS("Dumbbells"),
-  JUMPING_ROPE("JumpingRope"),
-  HORIZONTAL_BAR("HorizontalBar"),
-  BARS("Bars");
-
-  companion object {
-    fun byNameIgnoreCaseOrNull(input: String?): WktEquipment? {
-      return WktEquipment.values().firstOrNull { it.name.equals(input, true) }
-    }
-  }
-}
