+++ ok-workoutapp-common/src/commonMain/kotlin/models/WktFeedbackId.kt
@@ -1,4 +1,4 @@
-package models
+package ru.otus.otuskotlin.workoutapp.common.models
 import kotlin.jvm.JvmInline
 
 @JvmInline
