--- ok-workoutapp-common/src/commonMain/kotlin/models/WktFeedbackUser.kt
@@ -1,6 +1,0 @@
-package models
-
-data class WktFeedbackUser(
-  var id: WktUserId = WktUserId.NONE,
-  var name: String = "",
-)
