+++ ok-workoutapp-common/src/commonMain/kotlin/models/WktState.kt
@@ -1,4 +1,4 @@
-package models
+package ru.otus.otuskotlin.workoutapp.common.models
 
 enum class WktState {
   NONE,
