--- ok-workoutapp-common/src/commonMain/kotlin/models/WktWorkoutSearchGroupBy.kt
@@ -1,7 +1,0 @@
-package models
-
-enum class WktWorkoutSearchGroupBy(s: String) {
-  NONE("NONE"),
-  EQUIPMENT("EQUIPMENT"),
-  WORKOUT_TYPE("WORKOUT_TYPE")
-}
