--- ok-workoutapp-common/src/commonMain/kotlin/models/WktWorkoutType.kt
@@ -1,8 +1,0 @@
-package models
-
-enum class WktWorkoutType(s: String) {
-  NONE("NONE"),
-  CORE("Core"),
-  ARMS("Arms"),
-  LEGS("Legs");
-}
