+++ ok-workoutapp-feedback-common/src/commonMain/kotlin/models/WktFeedback.kt
@@ -1,4 +1,5 @@
-package models
+package ru.otus.otuskotlin.workoutapp.feedback.common.models
+import ru.otus.otuskotlin.workoutapp.common.models.*
 
 data class WktFeedback(
   var id: WktFeedbackId = WktFeedbackId.NONE,
