+++ ok-workoutapp-feedback-common/src/commonMain/kotlin/models/WktFeedbackUser.kt
@@ -1,0 +1,8 @@
+package ru.otus.otuskotlin.workoutapp.feedback.common.models
+
+import ru.otus.otuskotlin.workoutapp.common.models.WktUserId
+
+data class WktFeedbackUser(
+  var id: WktUserId = WktUserId.NONE,
+  var name: String = "",
+)
