+++ ok-workoutapp-mappers-v1/build.gradle.kts
@@ -9,6 +9,7 @@ dependencies {
   implementation(kotlin("stdlib"))
   implementation(project(":ok-workoutapp-api-v1-jackson"))
   implementation(project(":ok-workoutapp-common"))
-
+  implementation(project(":ok-workoutapp-workout-common"))
+  implementation(project(":ok-workoutapp-feedback-common"))
   testImplementation(kotlin("test-junit"))
 }
