+++ ok-workoutapp-mappers-v1/src/main/kotlin/MappersV1ToTransportWorkout.kt
@@ -1,62 +1,37 @@
-import models.*
 import ru.otus.otuskotlin.workoutapp.mappers.v1.exceptions.UnknownWktCommand
 import ru.otus.otuskotlin.workoutapp.api.v1.models.*
+import ru.otus.otuskotlin.workoutapp.common.models.WktCommand
+import ru.otus.otuskotlin.workoutapp.common.models.WktState
+import ru.otus.otuskotlin.workoutapp.workout.common.models.*
 
-fun WktContext.toTransport(): Response = when (val cmd = command) {
-  WktCommand.FEEDBACK_CREATE -> toTransportFeedbackCreate()
-  WktCommand.FEEDBACK_READ -> toTransportFeedbackRead()
-  WktCommand.FEEDBACK_UPDATE -> toTransportFeedbackUpdate()
-  WktCommand.FEEDBACK_DELETE -> toTransportFeedbackDelete()
+fun WktWorkoutContext.toTransport(): Response = when (val cmd = command) {
   WktCommand.WORKOUT_SEARCH -> toTransportWorkoutSearch()
   WktCommand.WORKOUT_CREATE -> toTransportWorkoutCreate()
   WktCommand.WORKOUT_READ -> toTransportWorkoutRead()
   WktCommand.WORKOUT_UPDATE -> toTransportWorkoutUpdate()
-  WktCommand.NONE -> throw UnknownWktCommand(cmd)
+  else -> throw UnknownWktCommand(cmd)
 }
 
-fun WktContext.toTransportFeedbackCreate() = FeedbackCreateResponse(
-  requestId = this.requestId.asString().takeIf { it.isNotBlank() },
-  result = if (state == WktState.RUNNING) Result.SUCCESS else Result.ERROR,
-  errors = errors.toTransportErrors(),
-  feedback = this.feedbackCreateResponse.toTransportFeedback()
-)
-fun WktContext.toTransportFeedbackRead() = FeedbackReadResponse(
-  requestId = this.requestId.asString().takeIf { it.isNotBlank() },
-  result = if (state == WktState.RUNNING) Result.SUCCESS else Result.ERROR,
-  errors = errors.toTransportErrors(),
-  feedback = this.feedbackReadResponse.map { it.toTransportFeedback() }
-)
-fun WktContext.toTransportFeedbackUpdate() = FeedbackUpdateResponse(
-  requestId = this.requestId.asString().takeIf { it.isNotBlank() },
-  result = if (state == WktState.RUNNING) Result.SUCCESS else Result.ERROR,
-  errors = errors.toTransportErrors(),
-  feedback = this.feedbackUpdateResponse.toTransportFeedback()
-)
-fun WktContext.toTransportFeedbackDelete() = FeedbackDeleteResponse(
-  requestId = this.requestId.asString().takeIf { it.isNotBlank() },
-  result = if (state == WktState.RUNNING) Result.SUCCESS else Result.ERROR,
-  errors = errors.toTransportErrors(),
-  feedback = this.feedbackDeleteResponse.toTransport()
-)
-fun WktContext.toTransportWorkoutSearch() = WorkoutSearchResponse(
+
+fun WktWorkoutContext.toTransportWorkoutSearch() = WorkoutSearchResponse(
   requestId = this.requestId.asString().takeIf { it.isNotBlank() },
   result = if (state == WktState.RUNNING) Result.SUCCESS else Result.ERROR,
   errors = errors.toTransportErrors(),
   workout = this.workoutSearchResponse.toTransport()
 )
-fun WktContext.toTransportWorkoutCreate() = WorkoutCreateResponse(
+fun WktWorkoutContext.toTransportWorkoutCreate() = WorkoutCreateResponse(
   requestId = this.requestId.asString().takeIf { it.isNotBlank() },
   result = if (state == WktState.RUNNING) Result.SUCCESS else Result.ERROR,
   errors = errors.toTransportErrors(),
   workout = this.workoutCreateResponse.toTransportWorkout()
 )
-fun WktContext.toTransportWorkoutRead() = WorkoutReadResponse(
+fun WktWorkoutContext.toTransportWorkoutRead() = WorkoutReadResponse(
   requestId = this.requestId.asString().takeIf { it.isNotBlank() },
   result = if (state == WktState.RUNNING) Result.SUCCESS else Result.ERROR,
   errors = errors.toTransportErrors(),
   workout = this.workoutReadResponse.toTransportWorkout()
 )
-fun WktContext.toTransportWorkoutUpdate() = WorkoutUpdateResponse(
+fun WktWorkoutContext.toTransportWorkoutUpdate() = WorkoutUpdateResponse(
   requestId = this.requestId.asString().takeIf { it.isNotBlank() },
   result = if (state == WktState.RUNNING) Result.SUCCESS else Result.ERROR,
   errors = errors.toTransportErrors(),
@@ -70,7 +45,6 @@ private fun WktWorkout.toTransportWorkout() = Workout(
   type = this.type.toTransport(),
   equipment = this.equipment.toTransport(),
   content = this.content.toTransport(),
-  feedback = this.feedback.map { it.toTransportFeedback() },
   rating = this.rating
 )
 
@@ -101,38 +75,10 @@ private fun WktWorkoutStep.toTransport() = WorkoutStep(
   timecode = this.timeCode
 )
 
-private fun WktFeedback.toTransportFeedback() = Feedback(
-  id = this.id.toString(),
-  review = this.review,
-  rating = this.rating,
-  user = this.user?.toTransport()
-)
-
-private fun WktFeedbackUser.toTransport() = FeedbackUser(
-  id = this.id.toString(),
-  name = this.name
-)
-
-private fun WktFeedbackPayload.toTransport() = FeedbackDeleteResponsePayload(
-  id = this.id.toString(),
-)
-
 private fun WktWorkoutSearchPayload.toTransport() = WorkoutSearchResponsePayload(
   groups = this.groups.map { it.toTransport() }
 )
 private fun WktWorkoutSearchResult.toTransport() = WorkoutSearchResponsePayloadGroupsInner(
   groupName = this.groupName,
   workouts = this.workouts.map { it.toTransportWorkout() }
-)
-
-private fun List<WktError>.toTransportErrors(): List<Error>? = this
-  .map { it.toTransport() }
-  .toList()
-  .takeIf { it.isNotEmpty() }
-
-private fun WktError.toTransport() = Error(
-  code = code.takeIf { it.isNotBlank() },
-  group = group.takeIf { it.isNotBlank() },
-  field = field.takeIf { it.isNotBlank() },
-  message = message.takeIf { it.isNotBlank() },
 )
