+++ ok-workoutapp-mappers-v1/src/main/kotlin/exceptions/UnknownWktCommand.kt
@@ -1,5 +1,5 @@
 package ru.otus.otuskotlin.workoutapp.mappers.v1.exceptions
 
-import models.WktCommand
+import ru.otus.otuskotlin.workoutapp.common.models.WktCommand
 
 class UnknownWktCommand(command: WktCommand) : Throwable("Wrong command $command at mapping toTransport stage")
