+++ ok-workoutapp-workout-common/src/commonMain/kotlin/models/WktEquipment.kt
@@ -1,0 +1,10 @@
+package ru.otus.otuskotlin.workoutapp.workout.common.models
+
+enum class WktEquipment {
+  NONE,
+  OWN_WEIGHT,
+  DUMBBELLS,
+  JUMPING_ROPE,
+  HORIZONTAL_BAR,
+  BARS
+}
