+++ ok-workoutapp-workout-common/src/commonMain/kotlin/models/WktWorkout.kt
@@ -1,4 +1,6 @@
-package models
+package ru.otus.otuskotlin.workoutapp.workout.common.models
+
+import ru.otus.otuskotlin.workoutapp.common.models.WktWorkoutId
 
 data class WktWorkout (
   var id: WktWorkoutId = WktWorkoutId.NONE,
@@ -7,6 +9,5 @@ data class WktWorkout (
   var type: WktWorkoutType = WktWorkoutType.NONE,
   var equipment: WktEquipment = WktEquipment.NONE,
   var content: WktWorkoutContent = WktWorkoutContent(),
-  var feedback: MutableList<WktFeedback> = mutableListOf(),
   var rating: Double = 0.0
 )
