+++ ok-workoutapp-workout-common/src/commonMain/kotlin/models/WktWorkoutContent.kt
@@ -1,4 +1,4 @@
-package models
+package ru.otus.otuskotlin.workoutapp.workout.common.models
 
 data class WktWorkoutContent(
   var video: String = "",
