+++ ok-workoutapp-workout-common/src/commonMain/kotlin/models/WktWorkoutSearchGroupBy.kt
@@ -1,0 +1,7 @@
+package ru.otus.otuskotlin.workoutapp.workout.common.models
+
+enum class WktWorkoutSearchGroupBy {
+  NONE,
+  EQUIPMENT,
+  WORKOUT_TYPE
+}
