+++ ok-workoutapp-workout-common/src/commonMain/kotlin/models/WktWorkoutSearchResult.kt
@@ -1,4 +1,4 @@
-package models
+package ru.otus.otuskotlin.workoutapp.workout.common.models
 
 data class WktWorkoutSearchResult(
   var groupName: String = "",
