+++ ok-workoutapp-workout-common/src/commonMain/kotlin/models/WktWorkoutType.kt
@@ -1,0 +1,8 @@
+package ru.otus.otuskotlin.workoutapp.workout.common.models
+
+enum class WktWorkoutType {
+  NONE,
+  CORE,
+  ARMS,
+  LEGS
+}
