+++ settings.gradle.kts
@@ -21,4 +21,6 @@ include("m1l5-coroutines")
 include("m2l3-testing")
 include("ok-workoutapp-api-v1-jackson")
 include("ok-workoutapp-common")
+include("ok-workoutapp-feedback-common")
+include("ok-workoutapp-workout-common")
 include("ok-workoutapp-mappers-v1")
